+++ Assets/_Source/Entities/Novatar/NovatarConfig.cs
@@ -1,9 +1,9 @@
 ﻿using _Source.App;
+using _Source.Features.ActorBehaviours.Sensors.Data;
 using _Source.Features.Movement.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using _Source.Features.ActorBehaviours.Sensors.Data;
 using UnityEngine;
 
 namespace _Source.Entities.Novatar
@@ -33,8 +33,8 @@ namespace _Source.Entities.Novatar
         public RangeSensorConfig RangeSensorConfig => _rangeSensorConfig;
 
         [Header("Damage")]
-        [SerializeField] private double _touchDamage;
-        public double TouchDamage => _touchDamage;
+        [SerializeField] private int _touchDamage;
+        public int TouchDamage => _touchDamage;
 
         [Header("Visuals")]
         [SerializeField] private List<RelationshipVisualsConfig> _relationshipVisualsConfigs;
